% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dem_check}
\alias{dem_check}
\title{DEM input file status check}
\usage{
dem_check(file, .verbose = TRUE, ...)
}
\arguments{
\item{file}{XML file download from fgd}

\item{.verbose}{\code{logical}. suppress info input XML file's about DEM information.}

\item{...}{Additional arguments passed on to other functions.}
}
\description{
DEM input file status check
}
