% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{fdg_line_parse}
\alias{fdg_line_parse}
\title{Line element parsed}
\usage{
fdg_line_parse(file)
}
\arguments{
\item{file}{XML file download from fgd}
}
\description{
Line element parsed
}
\details{
type AdmArea, BldA, WA
}
