% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fdg_dem.R
\name{read_fgd_dem}
\alias{read_fgd_dem}
\title{Read and Parse Fundamental Geospatial Data (FGD) dem file}
\usage{
read_fgd_dem(
  file,
  resolution = c(5, 10),
  return_class = c("df", "raster", "stars")
)
}
\arguments{
\item{file}{Path to XML file}

\item{resolution}{the number of dem mesh size resolution: 5m or 10m}

\item{return_class}{one of return object class: 'df' (data.frame, default), 'raster' or 'stars'}
}
\description{
The JPGIS (GML) format file provided by FGD as input,
the digital elevation models in the file are read as a data.frame or spatial object (raster or stars).
Supporting FGD Version 4.1 (2016/10/31)
}
\seealso{
\url{https://fgd.gsi.go.jp/download/ref_dem.html}
}
