% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fdg_dem.R
\name{read_fgd_dem}
\alias{read_fgd_dem}
\title{Read and Parse FGD's XML dem file}
\usage{
read_fgd_dem(file, resolution = c(5, 10), return_class = c("df",
  "raster"))
}
\arguments{
\item{file}{Path to XML file}

\item{resolution}{the number of dem mesh size resolution: 5m or 10m}

\item{return_class}{one of return object class: 'df' (data.frame), 'raster'}
}
\description{
Supporting FGD Version 4.1 (2016/10/31)
}
